+++ src/com/kaylerrenslow/armaDialogCreator/main/HelpUrls.java
@@ -20,4 +20,5 @@ public interface HelpUrls {
 	String PROJECT_INIT_WINDOW = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Create,-Open,-and-Export-Projects";
 	String CONFIGURE_DIRECTORIES_POPUP = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Configuring-Directories";
 	String NEW_CONTROL_POPUP = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Creating-a-New-Control";
+	String EXPORT = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Create,-Open,-and-Export-Projects#exporting-a-project";
 }