+++ src/com/kaylerrenslow/armaDialogCreator/main/lang/FXControlLang.java
@@ -52,5 +52,9 @@ public interface FXControlLang {
 			}
 		}
 	}
+
+	interface FileChooserPane {
+		String LOCATE = "Locate";
+	}
 }